refactor(e2e): migrate collection fixtures to entity folder and po/
Moves collection seed/db into fixtures/collection/, CollectionPO into
fixtures/po/, and updates collection test imports to ~e2e paths.
