+++ apps/studio/tests/e2e/collection/collection-display-permissions.test.ts
@@ -1,15 +1,14 @@
 import { test } from "@playwright/test"
 import crypto from "crypto"
-import { provisionE2ESite } from "~e2e/fixtures/site"
-import { ensureUserOnboarded } from "~e2e/fixtures/user"
-import { RoleType } from "~prisma/generated/generatedEnums"
-
-import { TEST_EMAILS, roleTag } from "../fixtures/auth"
+import { TEST_EMAILS, roleTag } from "~e2e/fixtures/auth"
 import {
   createCollectionWithTagCategories,
   deleteCollection,
-} from "../fixtures/collection"
-import { openCollectionIndexEditor } from "../fixtures/helpers"
+} from "~e2e/fixtures/collection"
+import { openCollectionIndexEditor } from "~e2e/fixtures/helpers"
+import { provisionE2ESite } from "~e2e/fixtures/site"
+import { ensureUserOnboarded } from "~e2e/fixtures/user"
+import { RoleType } from "~prisma/generated/generatedEnums"
 
 let siteId: number
