+++ apps/studio/tests/e2e/collection/collection-display.test.ts
@@ -1,13 +1,12 @@
 import { expect, test } from "@playwright/test"
 import crypto from "crypto"
+import { TEST_EMAILS, roleTag } from "~e2e/fixtures/auth"
+import { getDraftIndexPage } from "~e2e/fixtures/collection"
+import { openCollectionIndexEditor } from "~e2e/fixtures/helpers"
 import { seedCollection } from "~e2e/fixtures/resource"
 import { provisionE2ESite } from "~e2e/fixtures/site"
 import { ensureUserOnboarded } from "~e2e/fixtures/user"
 import { RoleType } from "~prisma/generated/generatedEnums"
-
-import { TEST_EMAILS, roleTag } from "../fixtures/auth"
-import { getDraftIndexPage } from "../fixtures/collection.db"
-import { openCollectionIndexEditor } from "../fixtures/helpers"
 
 let siteId: number
 let indexPageId: string
