+++ apps/studio/tests/e2e/collection/collection-table.test.ts
@@ -1,12 +1,11 @@
 import { test } from "@playwright/test"
+import { TEST_EMAILS, roleTag } from "~e2e/fixtures/auth"
+import { seedCollectionPages } from "~e2e/fixtures/collection"
 import { DashboardPO } from "~e2e/fixtures/po"
 import { seedCollection } from "~e2e/fixtures/resource"
 import { provisionE2ESite } from "~e2e/fixtures/site"
 import { ensureUserOnboarded } from "~e2e/fixtures/user"
 import { RoleType } from "~prisma/generated/generatedEnums"
-
-import { TEST_EMAILS, roleTag } from "../fixtures/auth"
-import { seedCollectionPages } from "../fixtures/collection"
 
 let siteId: number
 let collectionId: string
