+++ apps/studio/tests/e2e/collection/collection-tags-permissions.test.ts
@@ -1,18 +1,17 @@
 import { expect, test } from "@playwright/test"
 import crypto from "crypto"
-import { PageEditorPO } from "~e2e/fixtures/po"
-import { getResourceDraftTagged } from "~e2e/fixtures/resource"
-import { provisionE2ESite } from "~e2e/fixtures/site"
-import { ensureUserOnboarded } from "~e2e/fixtures/user"
-import { RoleType } from "~prisma/generated/generatedEnums"
-
-import { TEST_EMAILS, roleTag } from "../fixtures/auth"
+import { TEST_EMAILS, roleTag } from "~e2e/fixtures/auth"
 import {
   createCollectionPage,
   createCollectionWithTagCategories,
   deleteCollection,
-} from "../fixtures/collection"
-import { CollectionPO } from "../fixtures/collection.po"
+} from "~e2e/fixtures/collection"
+import { PageEditorPO } from "~e2e/fixtures/po"
+import { CollectionPO } from "~e2e/fixtures/po"
+import { getResourceDraftTagged } from "~e2e/fixtures/resource"
+import { provisionE2ESite } from "~e2e/fixtures/site"
+import { ensureUserOnboarded } from "~e2e/fixtures/user"
+import { RoleType } from "~prisma/generated/generatedEnums"
 
 const TAG_OPTION_ID = crypto.randomUUID()
