+++ apps/studio/tests/e2e/collection/filters-permission.test.ts
@@ -1,15 +1,14 @@
 import { test } from "@playwright/test"
 import crypto from "crypto"
-import { provisionE2ESite } from "~e2e/fixtures/site"
-import { ensureUserOnboarded } from "~e2e/fixtures/user"
-import { RoleType } from "~prisma/generated/generatedEnums"
-
-import { roleTag, TEST_EMAILS } from "../fixtures/auth"
+import { roleTag, TEST_EMAILS } from "~e2e/fixtures/auth"
 import {
   createCollectionWithTagCategories,
   deleteCollection,
-} from "../fixtures/collection"
-import { CollectionPO } from "../fixtures/collection.po"
+} from "~e2e/fixtures/collection"
+import { CollectionPO } from "~e2e/fixtures/po"
+import { provisionE2ESite } from "~e2e/fixtures/site"
+import { ensureUserOnboarded } from "~e2e/fixtures/user"
+import { RoleType } from "~prisma/generated/generatedEnums"
 
 let siteId: number
