+++ apps/studio/tests/e2e/fixtures/collection/index.ts
@@ -1,0 +1,2 @@
+export * from "./db"
+export * from "./seed"
